Update article layout base component
