+++ src/components/atoms/ArticlesPreviewWrapper.tsx
@@ -4,4 +4,5 @@ export const ArticlesPreviewWrapper = styled.div`
   display: flex;
   flex-wrap: wrap;
   justify-content: space-between;
+  align-items: center;
 `;
